pages components changes
